pyproject.toml
[project]
name = "nhl-cartography"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.25.0",
    "google-generativeai>=0.8.5",
    "matplotlib>=3.10.3",
    "networkx>=3.5",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psutil>=7.0.0",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
]
